@@ -1,23 +1,38 @@
 {
-	"name": "pecee/dot-notation",
-	"type": "library",
-	"description": "Dot notation support for JSON-like syntax for easily creating arrays in PHP.",
-	"keywords": ["php", "array", "json", "syntax", "multidimensional"],
-	"homepage": "https://github.com/skipperbent/instagram-php-sdk",
-	"license": "MIT",
-	"authors": [
-		{
-			"name": "Simon Sessingø",
-			"email": "simon.sessingoe@gmail.com",
-			"homepage": "http://www.pecee.dk"
-		}
-	],
-	"require": {
-		"php": ">=5.3.0"
-	},
-	"autoload": {
-		"psr-4": {
-			"Pecee\\": "src/Pecee/"
-		}
-	}
+  "name": "pecee/dot-notation",
+  "type": "library",
+  "description": "Dot notation support for JSON-like syntax for easily creating and mapping arrays in PHP.",
+  "keywords": [
+    "php",
+    "array",
+    "mapping",
+    "array-convert",
+    "move-array-key",
+    "json",
+    "syntax",
+    "multidimensional"
+  ],
+  "homepage": "https://github.com/skipperbent/dot-notation",
+  "license": "MIT",
+  "support": {
+    "source": "https://github.com/skipperbent/dot-notation/issues"
+  },
+  "authors": [
+    {
+      "name": "Simon Sessingø",
+      "email": "simon.sessingoe@gmail.com",
+      "homepage": "http://www.pecee.dk"
+    }
+  ],
+  "require": {
+    "php": ">=5.4.0"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^6.0"
+  },
+  "autoload": {
+    "psr-4": {
+      "Pecee\\": "src/Pecee/"
+    }
+  }
 }
